Huzzah, we only fetch the project list when necessary
That's because `mut` in `mut x: &i32` is responsible for the
mutability of the variable x (i.e. it determines whether
you can reassign x to another reference), while `mut` in
`x: &mut i32` defines the mutability of the reference itself,
i.e. it determines if you can reassign the value this
reference points at.